feat: legacy notes + checkbox
This adds a note variant, which can be enabled using !LEGACY as the first text. It adds a blockquote using the "deprecated API" look.
Additionally, a checkbox is added to the bottom (where the TS toggle was previously) to optionally disable those contents.

The idea is to add legacy syntax inline into the documentation at the appropriate places using these notes, and devs can hide those when unchecking the checkbox. The "Reactivity Fundamentals" page shows how this would look like.