feat: legacy details
This adds a details blocks for legacy mode. If one is toggled, all are toggled. Alternative to #80

The idea is to add legacy syntax inline into the documentation at the appropriate places using these notes, and devs can hide those when unchecking the checkbox. The "Reactivity Fundamentals" page shows how this would look like.