Make anchor underlines react better to clicks
This ensures that the anchor that is clicked is always the one that gets
underlined, even if just scrolling to that heading would underline a
different anchor.

Using `onhashchange` alone instead of adding `onclick`s to each anchor
is not sufficient because it's possible to click on an anchor that
already has its hash in the URL but has been scrolled away from.